Update Docker to 18.09.6 in StarlingX
kmem accounting feature was backported by RedHat and seems to be
incomplete and prone to memory leaks (and possible deadlocks).
This issue is triggered by Docker container runtime that activates
the kmem accounting feature for any cgroup by default. As a result
kmem accounting in runc is disabled on RHEL/CentOS starting 18.19.1
Updating Docker version to 18.09.6 in StarlingX to fix the issue.

Note that the client and container runtime are in separate packages
in Docker Engine 18.09. So, adding containerd.io and docker-cli to
fulfill the depenndencies for Docker Engine.


Change-Id: Ib229eb7ac4db45dbdf1260c6505242147e06838f
Closes-bug: 1831485
Signed-off-by: Alex Kozyrev <alex.kozyrev@windriver.com>