Add DNF packages
Update the rpms_centos.lst file to include the DNF packages.

The patching framework currently uses the smart package manager for
managing in-release software updates, while CentOS 8 is using DNF. In
order to facilitate the upgrade to CentOS 8 and python3, the use of
smartpm by the patch-agent is being replaced with the python2 dnf
modules. This should allow for an easier transition to python3.

Change-Id: Ie9d2387158833e0f76504a6d764a69c209d01458
Story: 2006227
Task: 37933
Signed-off-by: Don Penney <don.penney@windriver.com>