Constrain more-itertools for build failure
A recent update to the more-itertools python module causes a failure
in the STX_BUILD_container_setup CENGN build job. This module dropped
python-2.7 support after its 5.0.0 release. The newest update causes a
failure due to code that does not work with 2.7.

This commit adds a builder-constraints.txt file that the Dockerfile
passes to the pip install command to constrain module versions during
setup of the build container, allowing us to constrain the
more-itertools version to 5.0.0, the last version to officially
support python 2.7.

Change-Id: I3432c204ecd7c4ddedd8a7dea14216d4ec31e0aa
Closes-Bug: 1859642
Signed-off-by: Don Penney <don.penney@windriver.com>