Get tpm/integrity source code from CentOS 8.1 kernel tarball
The CentOS 4.18.0-147.3.1 kernel is not the same as upstream 4.18
kernel, and cause tpm/integrity from upstream cannot pass build.
There are many patches need be cherry-picked or reverted in order
to pass build with CentOS 8.1 kernel.
So switch to get tpm/integrity code from CentOS 8.1 kernel tarball.
With it, we just need focus on STX's patch.

Story: 2007308
Task: 38950
Depends-On: https://review.opendev.org/720325

Change-Id: I8090bc1aff7d5fe194567520093738099464b572
Signed-off-by: Shuicheng Lin <shuicheng.lin@intel.com>