dl_tarball: Add support for creating kernel-rt SRPM
This is for downloading the linux-rt tarball from CentOS
and then use it to create the SRPM via the CentOS common tooling.

This is needed as CentOS does not currently create a kernel-rt.src.rpm
package.

We clone the CentOS rpm/kernel-rt which contains the kernel-rt.spec
and use the into_srpm.sh tool which then downloads the correct linux
tarball from CentOS and generates the .src.rpm.  Since we use the rpm
name in the .lst file it will only be re-created if the rpm is not in
the downloads directory.

Story: 2007308
Task: 39570
Change-Id: Ieb0443fae4431b78173ef3fa42933b3d795a0816
Signed-off-by: Saul Wold <sgw@linux.intel.com>