dl_tarball: Add support for creating rt-setup SRPM
This is for downloading the rt-setup tarball from CentOS
and then use it to create the SRPM via the CentOS common tooling.

This is needed as CentOS does not currently create a rt-setup.src.rpm
package.

We clone the CentOS rpm/rt-setup which contains the rt-setup.spec
and use the into_srpm.sh tool which then downloads the correct linux
tarball from CentOS and generates the .src.rpm.  Since we use the rpm
name in the .lst file it will only be re-created if the rpm is not in
the downloads directory.

Signed-off-by: Yong Fu <fuyong@neusoft.com>
Story: 2007308
Task: 39558
Change-Id: Ie43ee5c6b345bf338ea63b9c2a8bfc2cb1430387