Fix train source repo config
The StarlingX_CentOS-OpenStack-train.repo had a copy/paste error,
where the train source repo was misconfigured with the stein label.
This conflicted with the Stein repo config, causing the Train source
repo not to be mirrored properly. This, in turn, causes build failures
because updated src rpm references cannot be found.

Change-Id: I323ffbce037a183802302c372b0e9510757b58b1
Signed-off-by: Don Penney <don.penney@windriver.com>