Centos 8 repo fixes
Centos 8 mirrored on cengn were set up by myself a month
ahead of the first definition in stx-tools. The cengn mirror
used the path centos/8.xxxx.  The centos-8/8.xxxx definition
that eventually appeared in stx-tools has created a costly
duplicate.  I'd like to remove the duplicate eventually.
I'll never be able to get rid of the duplication if it is
enshrined in the stx-4.0 release.

The keys are added to avoid verbose warnings during mirroring
on cengn.

Change-Id: I427bffc7fe3f37262b8c767584a486152d8473e2
Signed-off-by: Scott Little <scott.little@windriver.com>