cve_policy_filter: Upstream changed status keys
The upstream seems to have added a new 'fixedIn' key instead of just the
'notFixedYet' keys to the pkg map, so we need to change the logic for
setting the correct status.

The assumption of allfixed = fixed is because there are multiple packages
and any one NotFixedYet will invalid the allfixed at the CVE level.

Do some code clean-up along the way by removing get_status()

Change-Id: Ib896655b8cf81af2a9531e87964d8e37f74da7c2
Signed-off-by: Saul Wold <sgw@linux.intel.com>