build-iso failure due to missing repodata
Build-iso started failing due to the lack of
repodata in $MY_REPO/centos-repo/rt/Binary/ .

The problem was traced back to generate-centos-repo.sh
and an unintended modification of the dest_dir global
variable.  The unintended modification was due to
the new function copy_with_backup failing to use
'local' on it's internal variables.

Closes-bug: 1901920
Change-Id: If1c3928518552327af9e01a77b287849a1e3981b
Signed-off-by: Scott Little <scott.little@windriver.com>