generate-local-repo.sh bad ref to root.git
A previous de-branding update inadvertently included a
line intended for the de-branding of cgcs-root to root.
That portion of the debranding effort has been put on indefinite
hold as the 'repo' tool lacks critical features to ease the
transition.

This update reverts the one line change to it's prior state.

Closes-big: 1904908
Change-Id: Ie027d0eb4025611aaf086cfe38fa60535de2865d
Signed-off-by: Scott Little <scott.little@windriver.com>