Update git and git-review in Dockerfile
Opendev recently upgraded to a newer version of gerrit.
That gerrit demands that we run a newer version of git than
is available from Centos 7.  This will prevent users from
posting code for review from within this build container.

This update compiles a newer version (2.29.2) of git from
downloaded source code, and adds it to the build image.

A newer version of git-review is downloaded from pypi,
and added to the build image.

A mechanism is provided to feed in a valid email address
to the idocker image's ~/.gitconfig by defining
MY_EMAIL=you@your-email-host in your 'localrc' prior to 'tb.sh create'

Closes-bug: 1905789
Change-Id: I28ed185ea477ba3a8e47e1dd11fd384816f72717
Signed-off-by: slittle1 <scott.little@windriver.com>