Fix selection of a release specific mock prototype
Commit https://review.opendev.org/c/starlingx/root/+/762700
was intended to include the code to select the release specific
mock config prototype.

    e.g. mock.cfg.centos7.proto

As delivered, it would always select the default.

    e.g. mock.cfg.proto

When combined with a change in
https://review.opendev.org/c/starlingx/root/+/762701
that prematurely set the default mock cfg prototype
to the Centos 8 version while leaving the build
container at Centos 7, we created a build failure.

Closes-Bug: 1906547
Signed-off-by: Scott Little <scott.little@windriver.com>
Change-Id: If484b1b942d3a4febe5dee0afa01aaf0c00a4a1b
Signed-off-by: Scott Little <scott.little@windriver.com>