Dockerfile: fail in "yum install" on missing packages
By default "yum install" ignores packages that can't be downloaded and
return 0 to the shell.

In a Dockerfile such falsely successful commands are cached by
"docker build", so that the next time we run "docker build" the
"yum install" is not even attempted:

  # this "succeeds" and gets cached during the first "docker build",
  # even if the URL returns an error (eg DNS error or HTTP 404).
  # During a second "docker build" this bringis back the FS layer
  # from the cache and doesn't attempt to re-download the package
  # from that URL
  RUN yum install python3 http://some/url.rpm
  ...

This patch avoids the problem in case CENGN mirror is down.

Change-Id: I65f2185ce8a1832f3fc1be4135b0a49653654b5b
Closes-Bug: 1912682
Signed-off-by: Davlet Panech <davlet.panech@windriver.com>