fix for tb.sh dies on rmdir /var/lib/mock
tb.sh create might fail to create the builder docker image.
Yum install of the mock package failed, but yum did not report
the failure because other packages in the instalation set succeeded.
A subsequent command in the dockerfile fails when it tries to
remove/relocate /var/lib/mock, but failes because it is not present.

The yum error reporting was corrected in a recent update.
But this does not address cached copies of old and broken
yum install steps that pre-date the fix.

The mock package is paricularly sensitive as it has cengn as
the only source, where as other packages have multiple sources.

One option is to force docker to not use the cache at all, which
is slow.

The second option is to change the docker file, placing the
yum command to install mock under a seperate docker RUN command.
The altered build instructions ensure that the docker cache
with the broken install can't be used.  While we are at it,
move the user/project customization steps as far down as possible
to improve cache usage.

This change implements both.

Closes-Bug: 1917901
Signed-off-by: Scott Little <scott.little@windriver.com>
Change-Id: I28041bb44af53384c00a750b7162c6c6808c4e2d