Parallel downloads
Download_mirror.sh takes 15 hours to download all the rpms and
tarballs required to build StarlingX into a fresh workspace.
It should be much faster than that.

Replace the current serial download algorithm with one that is
parallel.  I'll cap it at 8 parallel downloads for now.  I'm
a little worried about overwelming CENGN.  This is sufficient
to drop download times from 15 to 3 hours for a fresh workspace,
and 30 min to 5 min to refresh an existing workspace.

Closes-Bug: 1918477
Signed-off-by: Scott Little <scott.little@windriver.com>
Change-Id: I469b4fee3cb304fe2984aa697ce2dc6cec52e79e