fix bad flockflock url
download_mirror.sh fails due to a bad path containing
‘stx-tools/centos-mirror-tools/config/centos/flockflock’

The path is constructed, and the trigger is when an EOL is missing
from a centos_build_layer.cfg file, causing 'cat' to merge the last
line of the offending file with the first line of the next file.

Switch 'cat' to 'grep', which will always ensure an EOL is present.
Along the way, we can filter out empty lines and comments.

Closes-bug: 1926987
Signed-off-by: Scott Little <scott.little@windriver.com>
Change-Id: I2404b3415f0f3e2f395c2bcb7a527aa01a488f61