Add devtoolset-8 related packages
Linux kernel 5.10 requires a gcc version beyond 4.9 in order to build
it.  So to build it we use a third party toolchain called devtoolset-8
which contains gcc version 8.3.  We only use this toolchain for the
kernel and kernel modules, leaving the original gcc 4.8.5 in place for
everything else.

scl-utils-build is required by devtoolset-8.

Story: 2008921
Task: 42550

Signed-off-by: Jackie Huang <jackie.huang@windriver.com>
Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>

Change-Id: I74381b933925801d0254df11c4d6eea6b42153d9