Add build dependencies of kexec-tools
In preparation of updating kexec-tools for the v5.10 kernel, this
commit:

- Adds the build dependencies of kexec-tools to the distro layer,

- Removes the kexec source RPM file reference from the flock layer,
  given that kexec-tools will now be built in the distro layer, and

- Modifies the dl_tarball.sh script so that the kexec-tools source RPM
  is downloaded and the "kdump-anaconda-addon-003-29-g4c517c5.tar.gz"
  source code archive is extracted from the source RPM file.

  This is done because this tarball does not appear to be publicly
  available otherwise. The only reference to the commit ID 4c517c5 is on
  a Red Hat employee's personal Github account at:
  https://github.com/ryncsn/kdump-anaconda-addon/commits/rhel-7

Verification: kexec-tools was successfully built with this patch in a
monolithic StarlingX build environment. A potential failure of wget was
found to be gracefully handled by inserting a "false" command after the
line where wget is called, and, in a separate experiment, by replacing
the wget command line with "false". (Thanks to Jim Somerville for
pointing out the buggy failure path in an earlier revision of this
patch.)

Story: 2008921
Task: 43040

Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>
Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>
Change-Id: Id8df3435d8660070cce31b9f994b386b020ed6e6