compiler layer: Add cmake-2.8.12.2 as a dependency of dwarves
Commit with change ID Ied2400c2c0a7ada6f974ac3f4299d1bd22b4f231 broke
the build because of a missing dependency (cmake) of the newly added
dwarves package in the compiler layer. This occurred because the commit
in question had only been tested in a monolithic build environment.

This commit adds the missing dependency (cmake-2.8.12.2) required by the
dwarves package to the compiler layer in order to unbreak the build.

Verification: The dwarves package was confirmed to not build with only
the compiler layer active, without this commit. Afterwards, this patch
was applied and the dwarves package was successfully built with only the
compiler layer active.

Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>
Change-Id: I1f7c02190338715d9b6600109d7527af0d3fd941