stx tool: Only change the attribute of pkgbuilder directory
After the commit https://review.opendev.org/c/starlingx/tools/+/814104
merged, when a new project was created the /localdisk directory will
be empty.
In userenv script we will occurs the failure due to lack of permission
becasue the /localdisk attribute was root:root.
So that the base file structure won't be created when the developer
login the builder container.
It will block the repo creating and package compiling.
Drop the -R option for chown command to avoid to cost many time when
there may be a very large number of files under /localdisk.
Meanwhile we need to change the channel directory attribute for latd.

Only change the double root attribute for the pkgbuilder directory now.

Story: 2008846
Task: 43744

Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>
Change-Id: I98877d531b64afd906de2786fd184cd49b636f1c