Add docker-ce 3rd party repository
In a centos build, the Centos mirror can be created from
a third party packages like docker-ce. The debdownloader
script that is used to populate the Debian mirror does not
support this.

The debdownloader script uses the apt repository to download
the binaries that are specified by the developer. If the
debian package is not specified and if it doesn't know
where to get the package it is not downloaded.

In order to workaround this, add the docker-ce repository so
that it is enabled and specify which docker-ce and associated
dependencies to download. This is needed because since
the starlingx/config-files package tries to configure
the docker-ce package rather than the Debian docker.io
package.

Story: 2008846
Task: 43814

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I96f5d426fcf7389d7565be4f883d3123a0bf61da