base-bullseye.lst: Add dependencies for mtce components
Add dependencies for the following packages:
 - mtce-dbgsym
 - mtce-hostw-dbgsym
 - mtce-hostw
 - mtce-hwmon-dbgsym
 - mtce-hwmon
 - mtce-lmon-dbgsym
 - mtce-lmon
 - mtce-pmon-dbgsym
 - mtce-pmon
 - mtce
 - mtce-compute
 - mtce-control
 - mtce-storage
skip:
 - mtce-common-dev
 - mtce-dev

This allows us to continue integration on Debian, by satisfying
dependencies during iso creation.
If there is any package listed here that we'll create a StarlingX
package for, then it should be removed when adding to stx-std.lst.

Aligned to 2 whitespace notation as most of the packages do.

Story: 2009101
Task: 43732
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: I85d007851e658e5451be492cb99c1b6057e9b0c6