stx: add new environment variable OSTREE_OSNAME
The OSTREE_OSNAME is required for building ostree, initramfs-ostree and
other related packages. Set it as an environment variable, allowing the
ostree name to be modified in the stx.conf.

Set OSTREE_OSNAME in $build_environment of debbuilder.conf, which is
used to setup the building environment, and add new function set_environ_vars()
in debbuilder.py to replace the @OSTREE_OSNAME@ in $build_environment
with environment variable.

Story: 2008846
Task: 43907

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>
Change-Id: I4e1adf9d385ff82888146220932fd1451b557646