Add priority to repositories in base-image
With the plugin yum-plugin-priorities now it's possible
to add a priority for each repo through --repo-priority
on command line and cfg file as well.

This is need because of the usage of ceph mirror and to force the
images to use ceph packages on that repo.

Test plan:

Docker images build succeeded.

Closes-Bug: #1949518
Signed-off-by: Delfino Curado <delfinogomes.curadofilho@windriver.com>
Change-Id: I3e2e4df3d75f290789188875b089eab4ca49bbb0