Add dependencies for upstream repo
This allows us to continue integration on Debian, by satisfying
dependencies during iso creation.
If there is any package listed here that we'll create a StarlingX
package for, then it should be removed the same time the StarlingX
package is merged.

Sort the file.

keystone : Depends: dbconfig-common but it is not installable
Depends: python3-pastescript but it is not installable
Depends: q-text-as-data but it is not installable
Depends: uwsgi-plugin-apparmor but it is not installable
Depends: uwsgi-plugin-python3 (>= 2.0.18) but it is not installable
openstack-dashboard : Depends: libjs-jquery-cookie but it is not
installable
openstack-dashboard-apache : Depends: apache2 (>= 2.4) but it is not
installable
Depends: libapache2-mod-wsgi-py3 but it is not installable
python3-aodhclient : Depends: python3-osprofiler but it is not
installable
python3-django-horizon : Depends: python3-django (>= 2:2.2) but it is
not installable
Depends: python3-django-compressor but it is not installable
Depends: python3-django-debreach but it is not installable
Depends: python3-django-pyscss but it is not installable
Depends: python3-enmerkar but it is not installable
Depends: python3-oslo.policy but it is not installable
Depends: python3-oslo.upgradecheck but it is not installable
Depends: python3-osprofiler but it is not installable
Depends: python3-pymongo but it is not installable
Depends: python3-pyscss (>= 1.3.7) but it is not installable
Depends: python3-redis but it is not installable
Depends: python3-semantic-version but it is not installable
Depends: python3-swiftclient but it is not installable
Depends: python3-xstatic but it is not installable
Depends: python3-xstatic-angular but it is not installable
Depends: python3-xstatic-angular-bootstrap but it is not installable
Depends: python3-xstatic-angular-fileupload but it is not installable
Depends: python3-xstatic-angular-gettext but it is not installable
Depends: python3-xstatic-angular-lrdragndrop but it is not installable
Depends: python3-xstatic-angular-schema-form but it is not installable
Depends: python3-xstatic-angular-ui-router but it is not installable
Depends: python3-xstatic-bootstrap-datepicker but it is not installable
Depends: python3-xstatic-bootstrap-scss but it is not installable
Depends: python3-xstatic-bootswatch but it is not installable
Depends: python3-xstatic-d3 but it is not installable
Depends: python3-xstatic-font-awesome but it is not installable
Depends: python3-xstatic-hogan but it is not installable
Depends: python3-xstatic-jasmine but it is not installable
Depends: python3-xstatic-jquery (>= 1.12.4.1) but it is not installable
Depends: python3-xstatic-jquery-migrate but it is not installable
Depends: python3-xstatic-jquery-ui but it is not installable
Depends: python3-xstatic-jquery.quicksearch but it is not installable
Depends: python3-xstatic-jquery.tablesorter but it is not installable
Depends: python3-xstatic-jsencrypt but it is not installable
Depends: python3-xstatic-magic-search but it is not installable
Depends: python3-xstatic-mdi but it is not installable
Depends: python3-xstatic-objectpath but it is not installable
Depends: python3-xstatic-rickshaw but it is not installable
Depends: python3-xstatic-roboto-fontface but it is not installable
Depends: python3-xstatic-smart-table but it is not installable
Depends: python3-xstatic-spin but it is not installable
Depends: python3-xstatic-term.js but it is not installable
Depends: python3-xstatic-tv4 but it is not installable
python3-gnocchiclient : Depends: python3-ujson but it is not installable
python3-heatclient : Depends: python3-swiftclient but it is not
installable
python3-ironicclient : Depends: python3-appdirs (>= 1.3.0) but it is not
installable
python3-keystone : Depends: python3-flask but it is not installable
Depends: python3-flask-restful but it is not installable
Depends: python3-ldappool but it is not installable
Depends: python3-oslo.policy (>= 3.0.2) but it is not installable
Depends: python3-oslo.upgradecheck but it is not installable
Depends: python3-osprofiler but it is not installable
Depends: python3-pysaml2 (>= 5.0.0) but it is not installable
Depends: python3-scrypt but it is not installable
python3-neutronclient : Depends: python3-os-client-config but it is not
installable
python3-openstacksdk : Depends: python3-appdirs (>= 1.3.0) but it is not
installable
Depends: python3-deprecation but it is not installable
Depends: python3-munch but it is not installable
Depends: python3-requestsexceptions but it is not installable
python3-pankoclient : Depends: python3-osprofiler but it is not
installable

Story: 2009101
Task: 43732
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: I05e48ad9a87113b1dac434e69d889ed93f225a23