debian: Fix pam configuration issues
pam-config requires libpam-pwquality and libpam-ldap
to be installed in order for it to work properly. It
is not included by default on the image so install it.

This is a debian only change that will not affect
Centos build. Tested by building image and trying
to change password.

Story: 2009101
Task: 44006

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I0665a7137128c217be7f8544d7020fb7c7e5d110