Fix intermittent rdma-core build failures
Intermittent build failures are seen that always include rdma-core.
It appears that rdma-core is both built by StarlingX, and listed as
a pre-built binary, and the two are not entirely compatible.

It seems to depend on what jobs were submitted to the same build engine
prior to building rdma-core. If the pre-built rdma-core was pulled into
that mock environment previously, the new rdma-core won't build.

There is no need for the pre-built rdma-core.  It needs to be
removed.

Closes-Bug: 1951463
Signed-off-by: Scott Little <scott.little@windriver.com>
Change-Id: Id3f84d2d12c6ca5ee16222504edd4b9d8d6356bc