debian: Setup sysadmin user
Set up the sysadmin user to change the default password when they
first login. This is done to keep parity with the Centos install.
Also lock the root user as well.

Tested by building image and checking that user is prompted
to change the password when they first login.

Story: 2009101
Task: 44147

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I4ef5222a7dbd6bb61fcd9b599d1bb89cb1f734ea