stx tool: Drop the previous way to sync cengn mirror in prepare task
Now we set the default value for the CENGNURL as the commit
https://review.opendev.org/c/starlingx/tools/+/821535 mentioned.
Since the mirrors of the debian packages are not initialized yet in
cengn server, we will encounter 400 Bad Request result in prepare task.

Just drop the mirror work for prepare task now.
We will implement the detailed program to download the debian binary
and source packages to replace it later.
Or once the specific directory structure for debian packages in cengn
server was created, we can recover it with the more accurate path if
necessary.

Story: 2008862
Task: 44219

Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>
Change-Id: I0c02e52aa4febe5de0a267ec88c9243e191208a8