Lock platformdirs to version 2.4.0 to retain python 2.7 support
Platformdirs was upgraded to 2.4.1 from 2.4.0 over the holidays,
and this breaks our build that requires python 2.7 compatibility.
We'll stay with 2.4.0 which has worked for StarlingX for some time,
rather than reverting to 2.0.2 which was the last with official
python 2.7 support.

Closes-Bug: 1956054
Signed-off-by: Scott Little <scott.little@windriver.com>
(cherry picked from commit 7bde482bcbee0d740446d7484a59d312e73e769b)
Change-Id: Ibb700dadac300b4eb4c1c4cc58a29ff57767c26e