stx tool: Set the default value of MINIKUBECPUS to 6
If we implement this project on a multi-cores server, as the
default value of MINIKUBECPUS is too small, and the building
will take a long time.
To speedup the packages and kernel building, we increase the
cpu number for minikube cluster. So the compiling time of the
kernel and kernel module is acceptable.

Story: 2008862
Task: 44346

Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>
Change-Id: I4466ecff8315ab07f8c998b4189c878796a962c9