gitignore build artifacts
Setting up a minikube build environment and building packages
leaves some artifacts in the tools git, visible to
'git status'.

In future we may wish to revisit this issue.  Can the
files be placed somewhere outsside of the git filesystem?
Can the files be cleaned up?

For now any such artifacts should be ignored by git, so
as to not confuse the users.

Story: 2008862
Task: 44682

Signed-off-by: Scott Little <scott.little@windriver.com>
Change-Id: I13a98b2b4700b1603be3feb1c8db675a49976a8e