debian: Update installer with fs size parameters
LAT is unable to export the fs size parameters to the
installer (VSZ,BSZ,RSZ,FSZ). The LAT installer defaults
the size parameters to values that do not match those
specified in the base-bullseye.yaml file.

Therefore such values need to be specified on the install
command line in every menu of every pxeboot grub file.
Any change in these values would require all the grub
files to be updated. This is not desirable.

So instead, this update patches the installer with the
values currently set in the base-bullseye.yaml file.

Until a way is found to extract the values directly
from base-bullseye.yaml they are hard coded to match
with this update.

A change of them in base-bullseye.yaml will require an
equivalent change like this in base-initramfs-bullseye.yaml

Test Plan:
PASS: Verify build-pkgs and build-image
PASS: Verify PXE install without parms on grub command line.

Regression:
PASS: Verify USB install

Note: CentOS env is not affected by this update.

Co-authored-by: Eric MacDonald <Eric.Macdonald@windriver.com>
Co-authored-by: Hongxu Jia <Hongxu.Jia@windriver.com>

Story: 2009101
Task: 44703
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>
Change-Id: I0601732c2ac02169d1bdc462c51ab50fd6d4e5a2