Debian: add dev patching certificate to the ISO
Adds the 'enable-dev-patch' package to the ISO on Debian.
That package allows developer patches to be testable.

This is not meant as a permanent solution.

There is currently no way to differentiate between
a developer ISO and a formal build, so once that capability
is implemented in the build, this will be adapted.

Test Plan:
  Pass - Build a Debian ISO
  Pass - Verify the certificate is installed in /etc/pki/wrs
  Pass - Verify patch upload passes the certificate check

Co-Authored-By: Jessica Castelino <jessica.castelino@windriver.com>
Story: 2009969
Task: 44951
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: Ifaa54be5e1bcb3c3bfebd8e4ae3620fd8d8a8cfe