dl_tarball:  Upgrade ice comms ddp fw to version 1.3.35.0
ICE driver update to support the Intel Logan Beach NIC (E810-CQDA2T).
and that dictates the upgrade of ddp firmware to ice_comms-1.3.35.0,
which is included in 27_1.zip.

https://www.intel.com/content/www/us/en/download/15084/intel-ethernet-adapter-complete-driver-pack.html

Remove the extra extraction step because it is zip file that we
want, rather than zips of zips.

Verification:
Success Path
- run download_mirror.sh, verify that the ice_comms zip file is
  downloaded in downloads folder.

Story: 2009952
Task: 44895

Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>
Change-Id: I7561543ba0bce5d2fd82a0315f81ba3dc3f1ba0b