Debian: Add k8s 1.23.1 package to the ISO
This change adds kubernetes 1.23.1 binaries to the debian image.

Currently, image creation fails if both k8s 1.21.8 and 1.23.1 are
included in the image. It fails during dpkg unpack as both versions
contain same binary names and build server tries to extract them to
the same location. Only one version can be included at this moment
until the above issue is fixed. A bug is open to address this issue.

Note: We do not immediately require to support multiple k8s versions.
But it is definitely required in the longer run. So, we can merge this
first, fix the above mentioned issue and then include both(multiple)
versions to the image.

Test Plan:
PASS: Image builds successfully.
PASS: AIO-SX installation successful.

Depends-On: https://review.opendev.org/c/starlingx/integ/+/831343

Depends-On: https://review.opendev.org/c/starlingx/integ/+/844649

Story: 2009830
Task: 44638

Signed-off-by: Kaustubh Dhokte <kaustubh.dhokte@windriver.com>
Change-Id: I6af8793f57f5734b07f4681b5c267ceaca183830