Debian: Add Mellanox drivers to the initramfs
This commit includes the Mellanox OFED device drivers in the initramfs.
This is necessary for systems that have Mellanox network adapters as
their management interfaces.

Please note that the mlnx-ofed-kernel-modules Debian package includes a
number of source files in /usr/src/ofa_kernel/x86_64 as well, so this
commit will unfortunately increase the size of the initramfs more than
necessary until the Debian packaging for the Mellanox OFED drivers is
improved.

Verification:
- An ISO image was successfully built with this commit.
- The /initrd file in the ISO image was unpacked and the existence of
  the Mellanox OFED device drivers in the initial RAM file system image
  was confirmed.
- The ISO image was started up in a qemu/KVM-based virtual machine and a
  root shell was acquired in the initramfs by appending "instsh=1" to
  the kernel command line. In the root shell, it was confirmed that the
  mlx5_core kernel module can be loaded with "modprobe mlx5_core".
- To confirm that this commit is necessary, an ISO image built without
  this commit was acquired from a nightly build server, and it was noted
  that the mlx5_core kernel module cannot be loaded in the initramfs
  environment using the procedure in the previous step.

Story: 2009878
Task: 45213

Change-Id: I30349d7eef7fc16c55d88064df576536d2864206
Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>