debbuilder.conf: Disable the apt upgrade of chroot
The chroot upgrading is enabled by default which causes
STX packages are installed and will not be purged. This
will cause the below issue:
The stx package 'systemd' is built and its sub-package
'libudev1' is uploaded to deb-local-build, then kernel is
built and libudev1 in chroot is upgraded to the STX version
from deb-local-build, after kernel built done, the STX libudev1
is kept in chroot. At this time, if 'build -c -a' is launched
again and linux kernel is built before 'systemd', for the local
repository 'deb-local-build' is empty, the 'systemd' from debian
official repository will be fetched, at this time sbuild complains:
"libudev-dev : Depends: libudev1 (= 247.3-7) but 247.3-7.stx.5 is to be installed"

Test Plan:
Pass: $build-pkgs -c -p systemd
      $build-pkgs -c -p openldap
      $stx control enter --dockername pkgbuilder
      $sbuild-shell chroot:bullseye-amd64-<USER>
      $dpkg -l | grep stx
      There should be no stx packages installed

Story: 2008846
Task: 45127

Signed-off-by: hbai <haiqing.bai@windriver.com>
Change-Id: Ib44bb73afa867d7a0c340734df93bd2ac462403f